+++ eveuniverse/management/commands/eveuniverse_purge_data.py
@@ -3,9 +3,10 @@ import logging
 from django.core.management.base import BaseCommand
 from django.db import transaction
 
-from ... import __title__
-from ...models import EveUniverseBaseModel
-from ...utils import LoggerAddTag
+from eveuniverse import __title__
+from eveuniverse.models import EveUniverseBaseModel
+from eveuniverse.utils import LoggerAddTag
+
 from . import get_input
 
 logger = LoggerAddTag(logging.getLogger(__name__), __title__)