+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [1.3.1] - 2023-08-05
+
+### Added
+
+- Added support for Django 4.0
+
+### Changed
+
+- Refactored GL script to use matrix
+- Refactoring
+- Added pylint to tox
+
+### Fixed
+
+- `EveEntity.objects.fetch_by_names_esi()` will not fetch unknown names if some names are already known
+
 ## [1.3.0] - 2023-07-11
 
 ### Added